pyproject.toml
[tool.poetry]
name = "lpt-example"
version = "0.1.0"
description = "Доска объявлений"
authors = ["Aleksandr Kostyakov <79514164692@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
alembic = "^1.12.0"
pydantic = "^2.3.0"
sqlalchemy = {extras = ["asyncpg"], version = "^2.0.20"}
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
python-keycloak = "^3.3.0"
python-multipart = "^0.0.6"
fastapi-third-party-auth = "^0.1.2"
oic = "^1.6.1"
jwt = "^1.3.1"
python-jose = "^3.3.0"
asyncpg = "^0.28.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 90
multi_line_output = 3
skip_gitignore = true
skip_glob = ["**/migrations/*", "**/settings/*"]
src_paths = ["backend"]

[tool.flake8]
max-line-length = 90
count = true
extend-ignore = ['E501', 'E203', 'W503']
exclude = [
    '.git',
    '__pycache__',
    'frontend',
    '.vscode',
    '*.egg',
    '.venv']

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
exclude = ["scripts", "docs", "test"]
